+++ src/libs/TransactionRecoveryApi/constants/chain/testnet.ts
@@ -1,4 +1,4 @@
-import { EvmChain } from "../../types";
+import { EvmChain } from "../../../types";
 import { Network } from "@ethersproject/networks";
 
 export const rpcMap: Record<EvmChain, string> = {
@@ -7,6 +7,8 @@ export const rpcMap: Record<EvmChain, string> = {
   [EvmChain.MOONBEAM]: "https://rpc.api.moonbase.moonbeam.network",
   [EvmChain.AVALANCHE]: "https://api.avax-test.network/ext/bc/C/rpc",
   [EvmChain.ETHEREUM]: "https://ropsten.infura.io/v3/510b6d5b3c56497b8070626a54f565a9",
+  [EvmChain.AURORA]: "https://testnet.aurora.dev",
+  [EvmChain.BINANCE]: "https://data-seed-prebsc-1-s1.binance.org:8545"
 };
 
 export const networkInfo: Record<EvmChain, Network> = {
@@ -30,4 +32,12 @@ export const networkInfo: Record<EvmChain, Network> = {
     chainId: 3,
     name: EvmChain.ETHEREUM,
   },
+  [EvmChain.AURORA]: {
+    chainId: 1313161555,
+    name: EvmChain.AURORA
+  },
+  [EvmChain.BINANCE]: {
+    chainId: 97,
+    name: EvmChain.BINANCE
+  }
 };